Cleanup and memory optimizing C# day 4

+++ cs/InputReader.cs
@@ -2,5 +2,5 @@ namespace AOC2023;
 
 public static class InputReader
 {
-    public static string ReadDayInput(int day) => File.ReadAllText($"../input/day{day}.txt");
+    public static string ReadDayInput(byte day) => File.ReadAllText($"../input/day{day}.txt");
 }
